requirements_simple.txt
Flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
Pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
boto3==1.28.85
Werkzeug==2.3.7